+++ tests/test_sequence/test_sequence.h
@@ -1,7 +1,7 @@
 #ifndef TEST_SEQUENCE_H
 #define TEST_SEQUENCE_H
 
-int testSequence();
+size_t testSequence();
 
 #endif
